Fix bogus Iceberg stats conversion for timestamp with time zone
Before the change, for data like `2021-07-24 02:43:57.348000 UTC`,
something like `1982-08-03 22:09:01.696 +05:44` was returned.

+++ plugin/trino-iceberg/src/main/java/io/trino/plugin/iceberg/PartitionTable.java
@@ -305,7 +305,7 @@ public class PartitionTable
             return Math.multiplyExact((Long) value, PICOSECONDS_PER_MICROSECOND);
         }
         if (type instanceof Types.FloatType) {
-            return Float.floatToIntBits((Float) value);
+            return (long) Float.floatToIntBits((Float) value);
         }
         if (type instanceof Types.IntegerType || type instanceof Types.DateType) {
             return ((Integer) value).longValue();
